Rename isolated package for console tests
