+++ systemtests/src/test/java/io/enmasse/systemtest/isolated/web/ChromeConsoleTest.java
@@ -2,7 +2,7 @@
  * Copyright 2018, EnMasse authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.enmasse.systemtest.isolated.console;
+package io.enmasse.systemtest.isolated.web;
 
 import io.enmasse.address.model.AddressSpaceBuilder;
 import io.enmasse.systemtest.bases.isolated.ITestIsolatedStandard;
